Move backend stubs out of codegen
