+++ src/backend/cuda/CMakeLists.txt
@@ -49,7 +49,7 @@ if(TILELANG_USE_CUDA_STUBS)
   #   - cuModuleLoadData, cuLaunchKernel, cuMemsetD32_v2, etc.
   # These can be called directly without any wrapper macros.
   # ============================================================================
-  add_library(cuda_stub SHARED src/backend/cuda/codegen/stubs/cuda.cc)
+  add_library(cuda_stub SHARED src/backend/cuda/stubs/cuda.cc)
   target_include_directories(cuda_stub PRIVATE ${CUDAToolkit_INCLUDE_DIRS})
   # Export symbols with visibility="default" when building
   target_compile_definitions(cuda_stub PRIVATE TILELANG_CUDA_STUB_EXPORTS)
@@ -77,7 +77,7 @@ if(TILELANG_USE_CUDA_STUBS)
   # The stub exports a minimal set of CUDA Runtime API entrypoints used by TVM
   # and lazily loads libcudart at runtime on first API call.
   # ============================================================================
-  add_library(cudart_stub SHARED src/backend/cuda/codegen/stubs/cudart.cc)
+  add_library(cudart_stub SHARED src/backend/cuda/stubs/cudart.cc)
   target_include_directories(cudart_stub PRIVATE ${CUDAToolkit_INCLUDE_DIRS})
   target_compile_definitions(cudart_stub PRIVATE TILELANG_CUDART_STUB_EXPORTS
                                                  ${TILELANG_CUDA_TOOLKIT_VERSION_DEFINITIONS})
@@ -111,7 +111,7 @@ if(TILELANG_USE_CUDA_STUBS)
   # The stub exports a minimal set of NVRTC C API entrypoints used by TVM and
   # lazily loads libnvrtc at runtime on first API call.
   # ============================================================================
-  add_library(nvrtc_stub SHARED src/backend/cuda/codegen/stubs/nvrtc.cc)
+  add_library(nvrtc_stub SHARED src/backend/cuda/stubs/nvrtc.cc)
   target_include_directories(nvrtc_stub PRIVATE ${CUDAToolkit_INCLUDE_DIRS})
   target_compile_definitions(nvrtc_stub PRIVATE TILELANG_NVRTC_STUB_EXPORTS
                                                   ${TILELANG_CUDA_TOOLKIT_VERSION_DEFINITIONS})
