+++ src/backend/cuda/stubs/cuda.cc
